Extract the code that's reusable in other loader out
